Angle calibration function is mostly complete.
For manual calibration.  The function is calibrate_m2angle.  If we were to add automatic calibration in the future, we could place the robot on a line, tell it to spin, and evaluate its own change of position.  However, if the robot only spins 180, it might be confused about its position and thinks it spun 360.